+++ src/main/scala/org/constellation/collection/SeqUtils.scala
@@ -1,0 +1,11 @@
+package org.constellation.collection
+
+import org.constellation.schema.Id
+
+object SeqUtils {
+
+  implicit class PathSeqOps(val seq: Seq[Id]) {
+    def formatPath: String = seq.map(_.short).mkString(" -> ")
+  }
+
+}